requirements.txt
Django==5.2.7
djangorestframework==3.16.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
asgiref==3.8.1
redis==5.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
python-dateutil==2.8.2
decorator==5.1.1
requests==2.32.4
httpx==0.27.0
websockets==14.2
uvicorn==0.34.0
gunicorn==23.0.0
whitenoise==6.6.0
django-cors-headers==4.3.1
django-environ==0.10.0
dj-database-url==3.0.1
Pillow==11.3.0
django-allauth==0.61.1
PyJWT==2.10.1
django-sendgrid-v5==1.3.0
sendgrid==6.11.0
sentry-sdk[django]==2.39.0
google-generativeai==0.3.2
# Force dependency re-install
agora-token-builder==1.0.0
# Cache-busting comment to force re-install